pyproject.toml
[tool.poetry]
name = "anyquote"
version = "0.3.0"
description = ""
authors = ["hsn <hsn1919810@gmail.com>"]
readme = "README.md"
packages = [
    { include = "anyquote" },
]
[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.4.0"
requests = "^2.32.3"
fonttools = "^4.53.1"
qrcode = "^7.4.2"
twitter-api-client = "^0.10.22"
webdriver-manager = "^4.0.2"
playwright = "^1.47.0"
rich = "^13.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
